#657 created news pages and sidebar for each news type. Need to change into a component
