+++ src/assets/data/navigation.json
@@ -312,34 +312,31 @@
       }
     }
   },
-  "News": {
+  "AllNews": {
     "id": "6",
-    "label": "News",
-    "path": "News",
+    "label": "All News",
+    "path": "AllNews",
+    "surveyLink": "https://forms.office.com/Pages/ResponsePage.aspx?id=SNwgxlAdUkmLOd9NVNdNgksLeUGJ_vlDujipnPAHzqlUQjU5WVc3UkQxVUNHRUIyS1lQVlQ4N0lGWiQlQCN0PWcu",
     "submenu": {
-	"all": {
-            "id": "6.0",
-            "label": "All News",
-            "path": "",
-            "surveyLink": "https://forms.office.com/Pages/ResponsePage.aspx?id=SNwgxlAdUkmLOd9NVNdNgksLeUGJ_vlDujipnPAHzqlUQjU5WVc3UkQxVUNHRUIyS1lQVlQ4N0lGWiQlQCN0PWcu"
+	"All": {
+	    "id": "6.0",
+	    "label": "All News",
+	    "path": "AllNews"
 	},
-	"news": {
+	"News Articles": {
 	    "id": "6.1",
 	    "label": "News Articles",
-	    "path": "news",
-            "surveyLink": "https://forms.office.com/Pages/ResponsePage.aspx?id=SNwgxlAdUkmLOd9NVNdNgksLeUGJ_vlDujipnPAHzqlUQjU5WVc3UkQxVUNHRUIyS1lQVlQ4N0lGWiQlQCN0PWcu"
+	    "path": "News"
 	},
-	"blog": {
+	"Blogs": {
 	    "id": "6.2",
 	    "label": "Blogs",
-	    "path": "blog",
-            "surveyLink": "https://forms.office.com/Pages/ResponsePage.aspx?id=SNwgxlAdUkmLOd9NVNdNgksLeUGJ_vlDujipnPAHzqlUQjU5WVc3UkQxVUNHRUIyS1lQVlQ4N0lGWiQlQCN0PWcu"    
+	    "path": "Blogs"
 	},
-	"podcast": {
+	"Podcasts": {
 	    "id": "6.3",
 	    "label": "Podcasts",
-	    "path": "podcast",
-            "surveyLink": "https://forms.office.com/Pages/ResponsePage.aspx?id=SNwgxlAdUkmLOd9NVNdNgksLeUGJ_vlDujipnPAHzqlUQjU5WVc3UkQxVUNHRUIyS1lQVlQ4N0lGWiQlQCN0PWcu"
+	    "path": "Podcasts"
 	}
     }
   }
