+++ src/components/FooterModule.vue
@@ -11,7 +11,7 @@
       <div class="column">
         <h4 class="title is-5">Media</h4>
         <ul>
-          <li><router-link to="/Media/News">News</router-link></li>
+          <li><router-link to="/Media/News/AllNews">News</router-link></li>
           <li><router-link to="/Media/Events">Events</router-link></li>
           <li>
             <span class="icon-text">
