+++ src/components/HomePageSidebarModule.vue
@@ -2,7 +2,7 @@
   <aside class="column is-3">
     <div class="content">
       <h3 class="title cve-news-title-padding">
-        <router-link class="has-text-black" to="/Media/News">News</router-link>
+        <router-link class="has-text-black" to="/Media/News/AllNews">News</router-link>
       </h3>
       <div>
         <table class="table is-narrow" role="presentation">
@@ -25,7 +25,7 @@
                 </span>
               </td>
               <td>
-                <router-link :to="`/Media/News/item/${$sanitize(newsItem.newsType)}/${$sanitize(newsItem.itemUrl)}`">
+                <router-link :to="`/Media/News/item/${$sanitize(newsItem.newsType)}/${newsItem.itemUrl}`">
                   {{$sanitize(newsItem.title)}}
                 </router-link>
               </td>
@@ -84,7 +84,7 @@
         </div>
 
       <div class="cve-expansion-bar">
-        <router-link class="cve-expansion-button" to="/Media/News">More news</router-link>
+        <router-link class="cve-expansion-button" to="/Media/News/AllNews">More news</router-link>
       </div>
       </div>
       <h3 class="title cve-events-title-padding">
