+++ src/router/index.js
@@ -25,8 +25,11 @@ import CNARules from '@/views/ResourcesSupport/AllResources/CNARules.vue';
 import BoardCharter from '@/views/ResourcesSupport/AllResources/BoardCharter.vue';
 import PrivacyPolicy from '@/views/Legal/PrivacyPolicy.vue';
 import TermsOfUse from '@/views/Legal/TermsOfUse.vue';
-import News from '@/views/Media/News.vue';
-import NewsItem from '@/views/Media/NewsItem.vue';
+import AllNews from '@/views/Media/News/AllNews.vue';
+import News from '@/views/Media/News/News.vue';
+import Blogs from '@/views/Media/News/Blogs.vue';
+import Podcasts from '@/views/Media/News/Podcasts.vue';
+import NewsItem from '@/views/Media/News/NewsItem.vue';
 import Events from '@/views/Media/Events.vue';
 
 Vue.use(VueRouter);
@@ -239,20 +242,38 @@ const routes = [
     },
   },
   {
-    path: '/Media/News',
+    path: '/Media/News/AllNews',
+    name: 'AllNews',
+    component: AllNews,
+    meta: {
+      title: 'All News | CVE',
+    },
+    alias: '/Media/News/AllNews',
+  },
+  {
+    path: '/Media/News/News',
     name: 'News',
     component: News,
     meta: {
       title: 'News | CVE',
     },
-    alias: '/Media/News/all',
+    alias: '/Media/News/News',
   },
   {
-    path: '/Media/News/:newsType',
-    name: 'News',
-    component: News,
+    path: '/Media/News/Blogs',
+    name: 'Blogs',
+    component: Blogs,
     meta: {
-      title: 'News | CVE',
+      title: 'Blogs | CVE',
+    },
+    alias: '/Media/News/Blogs',
+  },
+  {
+    path: '/Media/News/Podcasts',
+    name: 'Podcasts',
+    component: Podcasts,
+    meta: {
+      title: 'Podcasts | CVE',
     },
   },
   {
