+++ src/views/Media/News/News.vue
@@ -1,0 +1,138 @@
+<template>
+  <div id="cve-secondary-page-main-container" class="container">
+    <div class="columns is-centered">
+      <div class="column is-8-desktop cve-main-column-content-width is-12-tablet">
+        <main id="cve-main-page-content" role="main">
+          <h1 :id="cvenavs['AllNews']['submenu']['News Articles']['id']" class="title">
+            {{cvenavs['AllNews']['submenu']['News Articles']['label']}}
+          </h1>
+          <p class="cve-help-text" style="padding-bottom:4%;">Many of the links on this page redirect to external websites
+            <span class="icon cve-icon-xxs help-text-icon"><font-awesome-icon icon="external-link-alt" aria-labelledby="extenalSiteLinks"/></span>;
+            these links will open a new window or tab depending on the web browser used.
+          </p>
+          <div class="content" v-for="newsItem in pagination.currentNewsToDisplay" :key="`${newsItem.id}-${newsItem.newsType}`">
+            <article class ="media">
+              <div class ="media-content">
+                <div class ="content">
+                  <h2 class="title mt-4 mb-1 is-size-4">
+                    <router-link :to="`/Media/News/item/${$sanitize(newsItem.newsType)}/${newsItem.itemUrl}`">
+                      {{$sanitize(newsItem.title)}}
+                    </router-link>
+                  </h2>
+                  <span class ="level is-mobile has-text-grey-dark mb-1">
+                    <div class ="level-left">
+                      <span class ="level-item tag is-rounded has-text-grey-dark is-capitalized">{{newsItem.newsType}}</span>
+                      <p class="level-item" v-if="newsItem.newsType == 'blog'">By {{newsItem.author.name}}
+                        <span class="pl-1" v-if="newsItem.author.organization.name !== 'CVE'">from</span>
+                        <span class="icon-text" v-if="newsItem.author.organization.name !== 'CVE'">
+                          <a class="pl-1" :href="newsItem.author.organization.url">
+                          {{newsItem.author.organization.name}}
+                          <span class="icon is-size-7 cve-icon-xxs">
+                            <p id="extenalLink1" class="is-hidden">external site</p>
+                            <font-awesome-icon icon="external-link-alt" aria-labelledby="extenalLink1" aria-hidden="false" focusable="false"/>
+                          </span>
+                          </a>
+                        </span>
+                        <span class="pl-3">|</span>
+                      </p>
+                      <time class ="level-item" :datetime="newsItem.date">{{formatDate(newsItem.date)}}</time>
+                    </div>
+                  </span>
+                   <p v-if="newsItem.newsType == 'blog' || newsItem.newsType == 'news'"
+                    v-html="$sanitize(newsItem.description[0].content, {allowedTags: ['a', 'img']})">
+                   </p>
+                  <div class="is-flex-desktop" style="justify-content: left;"  v-if="newsItem.newsType == 'podcast'">
+                    <figure class="media-left image is3by2">
+                      <iframe class="has-ratio" :src="newsItem.url" frameborder="0" allowfullscreen>
+                      </iframe>
+                    </figure>
+                    <p v-html="$sanitize(newsItem.description[0].content)"></p>
+                  </div>
+                </div>
+              </div>
+            </article>
+          </div>
+          <nav class="pagination is-centered" aria-label="pagination">
+            <button class="pagination-previous button cve-button cve-button-outline" v-bind:disabled="pagination.currentPage > 1 ? false : true"
+              @click=getPreviousPage>
+              <p id="leftAngleIcon" class="is-hidden">previous page button</p>
+              <font-awesome-icon icon="angle-left" aria-labelledby="leftAngleIcon" aria-hidden="false"/>
+            </button>
+            <button class ="pagination-next button cve-button cve-button-outline"
+              v-bind:disabled="pagination.currentPage >= pagination.totalPages ? true : false"
+              @click=getNextPage>
+              <p id="rightAngleIcon" class="is-hidden">next page button</p>
+              <font-awesome-icon icon="angle-right" aria-labelledby="rightAngleIcon" aria-hidden="false"/>
+            </button>
+            <ul class ="pagination-list">
+              <li v-for="(pageNumber, index) in pagination.pages" :key="index" class="mt-1">
+                <a v-bind:class="pageNumber == pagination.currentPage ? 'pagination-link cve-button cve-button-outline is-current' :
+                  'pagination-link cve-button cve-button-outline'" @click=updatePagination(pageNumber)
+                  v-bind:aria-label="'Goto page ' + pageNumber"
+                  v-bind:aria-current="(pageNumber == pagination.currentPage) ? 'page' : false">{{pageNumber}}</a>
+              </li>
+            </ul>
+          </nav>
+        </main>
+        <SurveyLinkComponent :surveyLink="cvenavs['AllNews']['surveyLink']" />
+      </div>
+      <div class="column is-3 is-hidden-touch">
+        <NavigationSidebar :nav="cvenavs['AllNews']" />
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+import NavigationSidebar from '@/components/NavigationSidebar.vue';
+import SurveyLinkComponent from '@/components/SurveyLinkComponent.vue';
+import newsData from '../../../assets/data/news.json';
+import newsMixins from '../../../mixins/newsMixins';
+
+export default {
+  name: 'News',
+  components: {
+    NavigationSidebar,
+    SurveyLinkComponent,
+  },
+  data() {
+    return {
+      newsList: newsData,
+      newsItemsForType: [],
+      pagination: {
+        totalItems: 0,
+        currentPage: 1,
+        pageSize: 10,
+        totalPages: 0,
+        startIndex: 0,
+        endIndex: 0,
+        maxPages: 10,
+        pages: [],
+        currentNewsToDisplay: [],
+      },
+    };
+  },
+  mixins: [newsMixins],
+  created() {
+    // initialize data
+    this.newsItemsForType = this.getNewsByTypeAndAddUrlDetail('news');
+    this.paginate();
+  },
+  props: {
+    cvenavs: {
+      type: Object,
+      required: true,
+    },
+  },
+};
+</script>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped lang="scss">
+@import '../../../assets/style/globals.scss';
+
+.is-active {
+  color: $black;
+}
+
+</style>
