+++ src/views/Media/News/NewsArchives.vue
@@ -1,0 +1,22 @@
+<template>
+  <div id="cve-secondary-page-main-container" class="container">
+    <div class="columns is-centered">
+      <div class="column is-9-desktop is-9-tablet">
+        <main id="cve-main-page-content" role="main">
+          <h1 class="title">
+            News Archives
+          </h1>
+	  <p>Intro paragraph for news archives</p>
+	  <h2>By Year</h2>
+	  <div class="content">
+	    <ul>
+	      <li v-for="archive in archivedNews" :key="archive.id">
+	        <a :href="archive.url">{{archive.year}}</a>
+              </li>
+	    </ul>
+	  </div>
+	</main>
+      </div>
+    </div>
+  </div>
+</template>
