+++ src/views/Media/News/NewsItem.vue
@@ -52,7 +52,7 @@
 
 <script>
 import newsData from '@/assets/data/news.json';
-import createShortTitle from '../../mixins/newsMixins';
+import createShortTitle from '../../../mixins/newsMixins';
 
 export default {
   name: 'NewsItem',
@@ -84,6 +84,6 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import '../../assets/style/globals.scss';
+@import '../../../assets/style/globals.scss';
 
 </style>
